Split embargo nodes/lifecycle.py to satisfy BTND-07-004 (#912)
Extract PersistEmbargoEventNode and CommitLogCascadeNode from
lifecycle.py (512 lines → 402 lines) into a new cascade.py module.
lifecycle.py was the only leaf module exceeding the 500-line limit.

Also migrate flat test_nodes.py to per-submodule test files under
test/core/behaviors/embargo/nodes/ (BTND-07-002 SHOULD), adding
dedicated coverage for the new cascade.py nodes.
